complete controller and fix save bug

+++ TripMs/src/main/java/tpe/tripms/dto/TripResponseDTO.java
@@ -7,8 +7,8 @@ import tpe.tripms.model.Trip;
 
 @Data
 public class TripResponseDTO {
-	private int idTrip;
-	private int idScooter;
+	private long idTrip;
+	private long idScooter;
 	private int idDestinationStop;
 	private int idOriginStop;
 	private Date startDate;
